removes unused state layer from the monad stack
The previous version of multipod used a state to remember the
synchronised podcasts. Since we are now using a database we are not
using it anymore. This patch removes the state monad from the stack. It
may be the case that I am not interacted with the database properly and
there will be need for it in the future again.