add yesod to lookup synced podcasts
It is now possible to lookup the list of synced podcast by opening a
browser on port 3000 of the machine.
It is still necessary to use the repl on the server to add podcasts to
the list.

The structure would likely be improved by the scaffolding that comes
with Yesod. A future task should make that happen.