add and add option to the web interface
The web interface was lacking a way to add new podcasts. This patch add
a form that takes an url and add the corresponding podcast to the
database.

The errors are not elegantly handled, this should be done later.
In the current state the repl is not needed anymore but is still
available.