remove the repl
Multipod interface used to be based on the repl. The point of the app is
to centralize the podcasts information to be able to keep track of the
podcasts on different devices. Therefore the only necessary interface
for this side of the app is the web one.

Developing a repl based client will be done in the future.