add an entity representing an episode in the database
Multipod used to parse the rss everytime a specific podcast page was
loaded. This patch creates an entity representing an episode. This is
used to avoid the unnecessary parsing.

Note that because of that new episodes will not show up. Therefore, an
update function should be added to avoid stale data.