Use ajax to interact with the episodes
The episode read/unread status handing used to be handled through the
form. Since the aim is for this website to also act as a server to
different remote clients it was meaningful to decouple the view from the
controller.

This patch replaces the use of the form with ajax requests performing
dedicated POST requests..