Start using TypeScript instead of plain JavaScript.
This commit does not modify the code in any way it just sets up the
infrastructure to allow the typescript snippet to be properly compiled
to javascript.

All julius files are renamed to be theseus files to note that they are
not plain typescript and a new directory is added for external
librairy type signatures.